# services/alert/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Alert Agent API",
        "description": "HTTP API for alert submission, prompt management, and WebSocket real-time alert broadcasting",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Alert Verification microservice endpoint"
        }
    ],
    "paths": {
...
        "/api/v1/verification/ondemand": {
            "post": {
                "tags": [
                    "verification"
                ],
                "summary": "Verify alert on demand",
                "description": "Async on-demand verification using category-based prompt lookup. Returns HTTP 202 with a correlationId immediately. VLM processing runs in the background and results are published to Kafka / Elasticsearch via the same sink as the Kafka pipeline.",
                "operationId": "verify_ondemand_api_v1_verification_ondemand_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OnDemandVerificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Verification request accepted for background processing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OnDemandAcceptedResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unknown category or invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Server-generated job ID allocation conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Bounded job registry capacity exhausted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/verification/ondemand/{correlation_id}": {
            "get": {
                "tags": [
                    "verification"
                ],
                "summary": "Get on-demand verification status",
                "description": "Return bounded process-local job status. Terminal records expire and are not durable across Alert Bridge restarts.",
                "operationId": "get_ondemand_status_api_v1_verification_ondemand__correlation_id__get",
                "parameters": [
                    {
                        "name": "correlation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Correlation Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retained job status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationJobStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid correlation ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found or terminal receipt expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "verification"
                ],
                "summary": "Cancel on-demand verification before sink publication",
                "description": "Cancellation is accepted only while queued or running. Once sink publication begins, cancellation is rejected to avoid a false guarantee that no event was delivered.",
                "operationId": "cancel_ondemand_api_v1_verification_ondemand__correlation_id__delete",
                "parameters": [
                    {
                        "name": "correlation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Correlation Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job was already cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationJobStatus"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Cancellation accepted before publication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationJobStatus"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid correlation ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found or terminal receipt expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Publication already started or job is terminal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationJobStatus"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "OnDemandAcceptedResponse": {
                "properties": {
                    "status": {"type": "string", "const": "accepted", "title": "Status"},
                    "correlationId": {"type": "string", "title": "Correlationid"},
                    "statusUrl": {"type": "string", "title": "Statusurl"},
                    "message": {"type": "string", "title": "Message"},
                    "timestamp": {"type": "string", "title": "Timestamp"}
                },
                "type": "object",
                "required": ["status", "correlationId", "statusUrl", "message", "timestamp"],
                "title": "OnDemandAcceptedResponse"
            },
            "OnDemandVerificationRequest": {
                "properties": {
                    "category": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Category"
                    },
                    "info": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Info",
                        "description": "Must contain media_urls (list) and media_type ('video'|'image')"
                    }
                },
                "additionalProperties": true,
                "type": "object",
                "required": [
                    "category",
                    "info"
                ],
                "title": "OnDemandVerificationRequest",
                "description": "Full Incident payload for on-demand verification.\n\nAccepts the same message structure that DirectMedia receives from Kafka.\n``category`` and ``info`` (with ``media_urls`` + ``media_type``) are required;\nall other Incident fields are optional and passed through to the response.",
                "example": {
                    "category": "collision",
                    "end": "2026-04-22T10:00:30Z",
                    "id": "incident-123",
                    "info": {
                        "media_type": "video",
                        "media_urls": [
                            "http://localhost:30888/vst/sim/media/incident.mp4"
                        ]
                    },
                    "sensorId": "sensor-01",
                    "timestamp": "2026-04-22T10:00:00Z"
                }
            },
            "VerificationErrorResponse": {
                "properties": {
                    "status": {"type": "string", "const": "error", "title": "Status"},
                    "error": {"type": "string", "title": "Error"},
                    "message": {"type": "string", "title": "Message"},
                    "timestamp": {"type": "string", "title": "Timestamp"}
                },
                "type": "object",
                "required": ["status", "error", "message", "timestamp"],
                "title": "VerificationErrorResponse"
            },
            "VerificationJobError": {
                "properties": {
                    "code": {"type": "string", "title": "Code"}
                },
                "type": "object",
                "required": ["code"],
                "title": "VerificationJobError"
            },
            "VerificationJobResult": {
                "properties": {
                    "processingOutcome": {
                        "type": "string",
                        "enum": ["verified", "verification_failed", "unknown"],
                        "title": "Processingoutcome"
                    },
                    "sinkDelivery": {"$ref": "#/components/schemas/VerificationSinkDelivery"},
                    "verdict": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Verdict"},
                    "verificationResponseCode": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Verificationresponsecode"},
                    "errorSource": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Errorsource"}
                },
                "type": "object",
                "required": ["processingOutcome", "sinkDelivery"],
                "title": "VerificationJobResult"
            },
            "VerificationJobStatus": {
                "properties": {
                    "correlationId": {"type": "string", "title": "Correlationid"},
                    "state": {
                        "type": "string",
                        "enum": ["queued", "running", "publishing", "completed", "failed", "cancelled"],
                        "title": "State"
                    },
                    "terminal": {"type": "boolean", "title": "Terminal"},
                    "createdAt": {"type": "string", "title": "Createdat"},
                    "updatedAt": {"type": "string", "title": "Updatedat"},
                    "result": {"anyOf": [{"$ref": "#/components/schemas/VerificationJobResult"}, {"type": "null"}]},
                    "error": {"anyOf": [{"$ref": "#/components/schemas/VerificationJobError"}, {"type": "null"}]},
                    "cancellationAccepted": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Cancellationaccepted"}
                },
                "type": "object",
                "required": ["correlationId", "state", "terminal", "createdAt", "updatedAt"],
                "title": "VerificationJobStatus"
            },
            "VerificationSinkDelivery": {
                "properties": {
                    "transport": {"type": "string", "enum": ["elastic", "kafka", "unknown"], "title": "Transport"},
                    "outcome": {"type": "string", "enum": ["acknowledged", "submitted_unconfirmed", "failed", "unconfirmed"], "title": "Outcome"},
                    "documentId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Documentid"},
                    "index": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Index"},
                    "topic": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Topic"},
                    "partition": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Partition"},
                    "offset": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Offset"}
                },
                "type": "object",
                "required": ["transport", "outcome"],
                "title": "VerificationSinkDelivery"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}